requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-chroma
langchain-community
langchain-ollama
chromadb
python-dotenv
pypdf
docx2txt
unstructured[md]
streamlit
requests
pydantic
